requirements.txt
streamlit>=1.35.0
firebase-admin>=6.5.0
requests>=2.31.0
Pillow>=10.0.0
pyzbar>=0.1.9
pytesseract>=0.3.10
